hun_law: basic structure parser, down to Article
Third time's the charm. I tried using Higher level subdivisions as actual
strucutraal elements, and building a tree, but the  not-really-formedness
of real-world law texts prohibited that. This shouldn't stop us from
exporting this data as structured XML though, only the internal
representation will be flat.

See the following tags for more info:
2018.10.23-structures-by-splitting
2018.10.23-structures-line-by-line