hun_law: parse down to subpoints
Some parts are admittedly not that pretty (alphabetic subpoints and the
wrap_up detection), but it works nonetheless.

Interestingly, the whole SubArticleElement class started out as a simple
function (extract_multiple_from_text), which accepted a subelement class
and a bunch of bool parameters. Refactoring the whole thing into its own
class turned out to be a great idea, as the rest of the code became non-
redundaant and way easier to read.