hun_law: MVP amendment support
Only text change of full structural elements (article and below) are
implemented for now. Insertions and small text changes are next.

I had to implement an MVP reference parsing too, just to be able to
determine the structure level of the amendment. Just in case you were
wondering why I committed the first few Reference classes in an 'amendment'
commit. Oh, the Reference part was not 100% tested.