hun_law: separate structure from line-parsing
The structure objects will be immutable to ensure that further operations
(such as amending, and keeping track of what is in force) work correctly.
This will also separate the mediocre parsing code from the rest of the
codebase.

The parsing code also became mostly functional, with a hint of OOP, but
mainly for grouping functions.