parser: delete amendment_parser
It was a good experiment, but this is not the way to do it. Amendments
have to be parsed linearly, because of the "instead of" mechanic. I plan
to implement parsing them at amendment-time, jsut before applying them
this will allow us to make use of more sophisticated reference parsing,
and also don't have ugly circular references