utils: store cached parse data more efficiently
As in 10x more efficiently, for a bit of runtime cost.
The way it's done is relying on GZIP, and using a differential encoding
of X coordinates of individual characters in IndentedLine. This is
efficient, because even if the diff between coordinates are something like
10.7272900001, or a similarly long float, it's mostly the same for a
specific character, so it will appear many many times, making it super
compressable.