grammar.lark: make the grammar less ambigous
Unfortunately, if you leave the grammar ambigous, two things will happen:
it will become super slow (as the number of possible interpretations go up
exponentially), or it will simply produce the wrong result (when lark
"helpfully" does not even compute most of the interpretations).

The main reason the language is ambigous is that we want to allow pretty
much any text, and "lazily" but exhaustively find interesting parts. You
simply cannot have an "or else don't care" term in EBNF, so right now we
approximate it with weird "okay okay, you can have NUMBERs, but only if
you're sure it's not some kind of ID, so have two WORDS in front of it".

Maybe I should patch Lark to have an "or else" construct somehow.