parsers/semantic: completely rolled semantic data into structure elements
This means that the whole Act has to be regenerated with different data
(due to the immutable nature of the type) to enhance it with data from the
grammar analysis, but everything using this data will be way prettier and
easier. See the diff of html output for an example.