output/json: fully use built-in JSON serialization
This creates an unfortunate binding between tool output format and internal
structures, and will not be tolerable in the future. However, this whole
thing is in experimental phase, and this coupling should be removed once
something like a final JSON format is decided.

Until then, this is a nice little debugging and cache tool.