Semantic parsing: put semantic metadata outside actual structure
This saves us the trouble of having to completely rebuild the data
structure when doing the semantic analysis, at the cost of having to filter
what we need from all outgoing references every time we need to build the
html output.

This is a small price to pay for the removed complexity.