Parsing: Support for simple substitution block amendment parsing
The most basic use-case, when a single block is replaced. Two new parsers,
one for the semantic part (so that we know what the expected IDs are, and
later what type of amendment we are doing), and the structural part to
replace the quoted lines blocks.

Added BlockAmendment structure element, and support for writing it in all
3 output formats.

Added some utility functions to existing elements, including some functions
to make ArticleParser more similar to SAEParser.

Current limitations are many, this is mainly a demo for now.