Grammar analysis: Add support for very basic insertion parsing
The grammar itself had to be refactored so that amendments don't use
compound references. This is so that we will have greater control over what
kind of references can occur in amendments (e.g. no lists)

The tests had to be fixed a bit, since refs are now parsed correctly in
these cases.

There is no difference yet between replacement and amendment.