structure: move next_identifier to Structure
This will be used for Reference "next to" fuunctionality, and Reference
(or its parsing for that matter) is not part of structure parsing.

"First header" was not moved, since it is still pretty much in the domain
of stucture parsing. Also, I did not know what to do about the Prefixed
Subpoint thing, and the prefix is needed for the first identifier.

TODO, I guess.