tests: add tests for structure classes
Now that there is nontrivial logic in these classes, it is way faster to
have tests at this level for the more complex things (like the "next id"
thing) than at a higher test level.

The structure classes themselves had to be modified so that they are
created more easily manually.