parsing: support numeric subpoints
A lot of things had to be rewritten for this feature, and I will not be
splitting this commit up afterwards. Deal with it. Things done:
- NoSubpointsError renamed to NoSubelementsError, because that's a more
  accurate name
- intro parsing moved from extract_multiple_from_text() to a separate
  function.
- Now multiple possible subelement parsers are tried, not just the first
  one. The previous algorithm caused numeric SUBpoints to be parsed as
  numeric POINTS. Now we find the first subelement header in the text, and
  hope that it's the one that's the "top-level".
- Added NumericPointParser
- Added a convoluted test to test these various logics
- QuotedBlockParser became more tightly integrated into this scheme
- QuotedBlockParser now accepts “ as a starting character, so it's now
  consistent with iterate_with_quote_level

To be completely honest, the parsing logic became a mess, and it should
probably be refactored into mor logical units. Maybe next time. At least
we now have a ton of tests to be sure it works at least.