parsers/structure: include space after header in HEADER_REGEX
Otherwise, we might match 'headers', which don't actually have a space
after them. In this case, the following assertion will fail, because header
prefix generation in structure.py always includes the space:

293 identifier = cls.extract_identifier(lines[0])
294 prefix = cls.PARSED_TYPE.header_prefix(identifier)
295 assert lines[0].content.startswith(prefix)

This should not cause a problem, because real documents always have a space
after the header, even in block amendments. If not, it's probably the PDF
parser's fault.