parser/structure: parse complex, structural block amendments
Now that we have a grammar for it, we just had to repurpose the ActParsers
body parsing to parse the BAs' quoted text instead.

A non-strict mode for parsing structural elements inside quotes was needed,
because the structural elements inside can have any kinds of numbering,
maybe not have empty lines, etc. This parsing mode would be very bad for
Act parsing, but for the unformatted block amendments, we don't have much
choice.

Also, the assumption that all children are the same for all SAEs is now not
true. Fortunately no code actually depended on this invariant.