structure: put semantic data into elements rather than in the Act
Putting them in the Act was just a horrible hack to avoid creating a whole
new object tree. However this OOB data may get out of sync if the tree is
modified, so it's not a very good design.

Now deleting a node will delete all semantic data, modifying it will also
be way easier.