strucure, semantics: simplified BlockAmendments
Block Amendments should define two cut points in the children of the
amended "thing": the part between these two cuts is deleted, and the new
children are inserted into the hole.

Of course there may be a case where the two cuts don't delete any children:
e.g. both of them fall between the same two original children. This is the
"clean" insertion case.

This can be represented with a single, optionally range-type position, and
all the other fields are redundant:
- 'replaces' is implied by the new position
- 'expected_type' can be obtained from the position reference
- 'expected_id_range' is the last not-None component of the reference

One of the side effects is that some high level structural information
(e.g. in which chapter the insertion should happen) is lost, but I'm pretty
sure it's useless anyway. If not, there's still a place for them in
StructuralReference.

This should result in way cleaner and more straightforward user code, as
the only task is to find the two cut points.

This also eliminated the need to support Type fields in dict2object, but
I'm not deleting that functionality, as it may become useful in the future.