parsing/structure: add new heuristic to wrap_up detection
The old code only looked for indentation when detecting wrap_up. This is
still the default for Act parsing.

For BlockAmendment parsing, a new heuristic is also added: if a line is not
justified, the next line is probably part of the wrap_up, otherwise it is
part of the line.

This is not perfect, but fixes 90% of the current problems in Block
Amendments, without breaking anything else.