pylintrc: disable several warnings
consider-using-f-string:
While this would be nice, I won't refactor the wole codebase just
because of a pylint warning.

unspecified-encoding:
The only encoding that's relevant is utf-8. It won't ever be different
(fingers crossed)

global-variable-not-assigned:
Sometimes I want to make sure that the variable is global, even when
reading from it.